src/claudecart/database/vector_manager.py
import os
from typing import Dict, List, Any, Optional, Union

import lancedb
from fastembed import TextEmbedding


class VectorManager:
    """
    Vector database manager for semantic product search in ClaudeCart.
    
    This class handles the creation, management, and querying of vector 
    embeddings for product descriptions, enabling semantic search and
    recommendation capabilities.
    """
    
    def __init__(
        self, 
        db_path: str = "vectorstore", 
        embedding_model: str = "BAAI/bge-small-en-v1.5"
    ):
        """
        Initialize the vector database manager.
        
        Args:
            db_path: Path to the vector database directory
            embedding_model: Name of the embedding model to use
        """
        self.db_path = db_path
        self._ensure_db_exists()
        
        # Initialize embedding model
        self.embedding_model = TextEmbedding(embedding_model)
        
        # Connect to LanceDB
        self.db = lancedb.connect(self.db_path)
        
    def _ensure_db_exists(self) -> None:
        """Ensure vector database directory exists."""
        if not os.path.exists(self.db_path):
            os.makedirs(self.db_path)
    
    def _embed_text(self, text: str) -> List[float]:
        """
        Generate embeddings for text.
        
        Args:
            text: Text to embed
            
        Returns:
            Embedding vector
        """
        embeddings = list(self.embedding_model.embed([text]))
        return embeddings[0].tolist()
    
    def index_product(self, product: Dict[str, Any]) -> None:
        """
        Index a product in the vector database.
        
        Args:
            product: Product information dictionary
        """
        # Function to be implemented
        pass
    
    def semantic_search(
        self, 
        query: str, 
        limit: int = 5, 
        filters: Optional[Dict[str, Any]] = None
    ) -> List[Dict[str, Any]]:
        """
        Perform semantic search for products.
        
        Args:
            query: Search query string
            limit: Maximum number of results
            filters: Additional filters to apply
            
        Returns:
            List of matching product dictionaries with similarity scores
        """
        # Function to be implemented
        pass
    
    def get_similar_products(
        self, 
        product_id: int, 
        limit: int = 5
    ) -> List[Dict[str, Any]]:
        """
        Find products similar to a given product.
        
        Args:
            product_id: ID of the reference product
            limit: Maximum number of similar products to return
            
        Returns:
            List of similar product dictionaries with similarity scores
        """
        # Function to be implemented
        pass